requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
jsonpickle==0.9.6
python-dateutil==2.7.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
